#18 style: Todo 리스트 불러오는 코드 위치 변경
